Removing unnecessary code
