--- backend/Infosafe_Backend/src/main/java/com/fragile/infosafe/service/PasswordEncryption.java
@@ -1,8 +1,0 @@
-package com.fragile.infosafe.service;
-
-import java.io.UnsupportedEncodingException;
-import java.security.GeneralSecurityException;
-
-public interface PasswordEncryption {
-    public String encryptPassword(String password) throws GeneralSecurityException, UnsupportedEncodingException;
-}
